refactor(lsp): extract roslyn diagnostics logic into separate module
- Move roslyn-specific diagnostic refresh setup to dedicated roslyn.lua module
- Wrap refresh logic in M.setup() function for cleaner initialization
- Keep main lsp init.lua focused on general LSP configuration
